Fix Dockerfile.fedora to use latest fedora releases and correct ENV and install nodemon in assemble for Fedora
Previosly, when nodemon was installed in the Dockerfile, it meant that the image
would be shipped with bits downloaded from the untrusted Internet source, which
is not ok for Fedora distribution and thus koji does not allow it at all (the
container build failed).

Moving it to the assemble script has a similar effect, but allows us to build
the image and we do not ship the bits in the container image.

A precedence of doing this is also in the python image that pulls PIP similarly.

Anyway, this should go away once we have nodemon properly packaged as an RPM:
https://bugzilla.redhat.com/show_bug.cgi?id=1433784