Make sure module is properly reset before attempt to install
Required-by: https://github.com/sclorg/s2i-base-container/pull/206

Reasoning: Reset is necessary for switching module streams. So, in case
there is already some nodejs module installed (thus enabled), we must
first call dnf module reset and only then enable it.
In case of version 14 and above, the Nodejs Dockefile is built on top
of s2i-core, not s2i-base, so we're sure there is no nodejs installed
in the parent image and don't need to run dnf module reset in that case.