Enable build&push for v20 instead v16 as it's released already
The goal to have cXs images is to have a testing image for the upcoming
version before it's released officially, so community can take a look
sooner. Since v16 and v18 are already released in the supported UBI
offering, there is no use case for the cXs based container images on
quay.io.